Crate rust_miscellaneous_pkg

A collection of useful functionalities I want to link it to other languages like R or Node.js

Modules

basic_arithmetic

Collection of basic arithmetic functions

basic_file_operation

Collection of basic file operations

bio_algorithms

This is a collection of bio algorithms

crypto

Collection of crypto algorithms

custom_request

Module to create HTTP basic Requests

docx_mod

excel_operations

This module allows to write a json object to an excel sheet and read out the content of excel sheet to json a json structure

faker

Collection to create different fake data

hashing

This a module with a collection of different hashing methods This is a wrapper for tiny_keccak

image_resize

This is a module which allows you to resize images. With this module it is possible to resize images with different algorithms.

machine_learning

matrix_operations

random_numbers

Module to create random numbers

regex_operations

Module for some regex operation

sql_operation

This is a module which allows you to get/insert/delete/update data from mysql/sqlite/postgres database

xml_to_json_convert

Module to convert json from xml