Module rust_miscellaneous_pkg::bio_algorithms::bio_algorithms

Functions

bndm_matching

This is a function which calculates the Backward nondeterministic matching

bom_matching

This is a function for the Backward oracle matching algorithm

hamming

This function calculates the hemming distance of two strings

hamming_smid

This calculates the SIMD-accelerated Hamming Distance

horspool_matching

This function which calculates the horspool matching

kmp_matching

This function which calculates the kmp matching

levenshtein

This function calculates the levenshtein distance of two strings

levenshtein_smid

This calculates the SIMD-accelerated Levenshtein